Add graphical phase shifts away from current level
Before, phase shifts could not occur /away from/ the current level. This
is because they would just disappear and that would not look good. So,
when monsters are in the current level they actually act out what "phase
shif"ing is supposed to be: if they run into a portal, they follow it.