Support nearest neighbor on most major browsers
Previously supported /exclusively/ firefox (which is horrible! and why
it wasn't published). Now it is known to support:
 - Firefox
 - Chrome
 - Firefox for iOS
 - Safari for iOS
I wouldn't be surpised if it supports all the top versions of the modern
browsers.

There is one downside to this all: webGL is out. It's canvas exclusively
now, speed be damned. This is worth it for the simplicity with which we
can now support all modern browsers.