Refactor portal interaction and support touch
Completely refactor portal interaction because it was a complete mess
and repeated everything twice for portalUp and portalDown. It now uses a
loop. Unfortunately it requires a bit of a hack in accessing portalHold
but it's not too bad. The code is much cleaner, and now it allows us to
make the neccessary change: you must hold the screen to move portal and
it won't shoot in reaction.

Also inadvertantly fixes bug that the portal charge would decrease too
fast if you were within the bounds.