Balance entropy and phase shifts
Overhaul of phase shift system: no longer recursive. Instead, if a
monster cannot be found around the level, nothing happens. Because all
levels are now constantly updating, this is not an issue. Eventually, a
monster will make it to you.

Also worked on making phase shifts happen the correct amount, which
involved balancing the amount of entropy you get for certain things.
Killing a monster to reduce entropy added urgency that made it a lot
more fun.